resize spatial param

+++ tests/test_resize.py
@@ -57,7 +57,7 @@ class TestResize(NumpyImageTestCase2D):
     )
     def test_correct_results(self, spatial_size, mode, anti_aliasing):
         """resize 'spatial_size' and 'mode'"""
-        resize = Resize(spatial_size, mode=mode, anti_aliasing=anti_aliasing)
+        resize = Resize(spatial_size, mode=mode, anti_aliasing=anti_aliasing, dtype=np.float64)
         _order = 0
         if mode.endswith("linear"):
             _order = 1
